Update Wordle ID type to ulong
The Wordle entity's ID type has been changed from int to ulong. This ID now can be used flexibly in different contexts such as guild, DM or API, with appropriate IDs like the channel or user IDs. Corresponding adjustments and a new database migration have been created to reflect this type change.

+++ Blink3.Core/Entities/Wordle.cs
@@ -14,9 +14,13 @@ public class Wordle
     /// <summary>
     ///     Gets or sets the ID of the Wordle Game.
     /// </summary>
+    /// <remarks>
+    ///     If used in a guild, this will be the channel ID
+    ///     In case it is used in DM or through the API, will be the user ID
+    /// </remarks>
     [Key]
     [Required]
-    public int Id { get; set; }
+    public ulong Id { get; set; }
 
     /// <summary>
     ///     Represents the word to guess in the Wordle game.
