+++ Blink3.Core/Repositories/Interfaces/IWordleRepository.cs
@@ -15,5 +15,5 @@ public interface IWordleRepository : IGenericRepository<Wordle>
     ///     A task representing the asynchronous operation.
     ///     The task result is a boolean indicating whether the entity exists or not.
     /// </returns>
-    public Task<bool> ExistsByIdAsync(int id);
+    public Task<bool> ExistsByIdAsync(ulong id);
 }
